Create commmon CodeView component for raw text
Since the search behaviour will be attached to an action associated with
the text view, in order to avoid repeating the same work I'll just
create a common component called CarteroCodeView and use it instead of
GtkSourceView.

Features of CarteroCodeView:

* Automatic dark theme when dark mode.
* Automatic unidirectional binding to accept the current settings.
