+++ data/ui/code_export_pane.blp
@@ -27,7 +27,7 @@ template $CarteroCodeExportPane: $CarteroBaseExportPane {
       vexpand: true;
       hexpand: true;
 
-      GtkSource.View view {
+      $CarteroCodeView view {
         styles [
           "use-cartero-font"
         ]
