+++ data/ui/raw_payload_pane.blp
@@ -23,7 +23,7 @@ template $CarteroRawPayloadPane: $CarteroBasePayloadPane {
     hexpand: true;
     vexpand: true;
 
-    GtkSource.View view {
+    $CarteroCodeView view {
       styles [
         "use-cartero-font"
       ]
