+++ data/ui/response_panel.blp
@@ -51,7 +51,7 @@ template $CarteroResponsePanel: Adw.Bin {
               hexpand: true;
               vexpand: true;
 
-              GtkSource.View response_body {
+              $CarteroCodeView response_body {
                 styles [
                   "use-cartero-font"
                 ]
